# api/collaboration/openapi.yaml
openapi: 3.1.0
info:
  title: AphelionDMM Collaboration Control API
  version: 1.0.0
  description: Version 1 lifecycle and snapshot contract. Filesystem paths and process commands are never accepted.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1
  - url: https://collaboration.example
paths:
...
  /v1/hosted/sessions/{session_id}/join:
    parameters: [{ $ref: '#/components/parameters/SessionID' }]
    post:
      summary: Resolve enabled membership or atomically admit a new active Community editor
      operationId: joinHostedSession
      security: [{ bearerToken: [] }]
      responses:
        '200':
          description: Existing roles and disabled restrictions take precedence; contains no credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedMembership' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /v1/hosted/sessions/{session_id}:
    parameters: [{ $ref: '#/components/parameters/SessionID' }]
    patch:
      summary: Replace owner-controlled labels and visibility without changing map revisions
      operationId: updateHostedSessionMetadata
      security: [{ bearerToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HostedSessionMetadata' }
      responses:
        '204': { description: Updated; existing admitted members retain access }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    launchToken: { type: http, scheme: bearer, description: Single-use embedded launch token }
    bearerToken: { type: http, scheme: bearer }
  parameters:
    SessionID:
      name: session_id
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
  responses:
    Error:
      description: Bounded stable error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    SHA256:
      type: string
      pattern: '^[0-9a-f]{64}$'
...
    HostedSessionMetadata:
      type: object
      additionalProperties: false
      properties:
        visibility: { type: string, enum: [private, community], default: private }
        title: { type: string, maxLength: 128, description: Maximum 128 UTF-8 bytes }
        map_label: { type: string, maxLength: 128 }
        environment_label: { type: string, maxLength: 128 }
...
    HostedMembership:
      type: object
      additionalProperties: false
      required: [session_id, actor_id, role]
      properties:
        session_id: { type: string, minLength: 1, maxLength: 128 }
        actor_id: { type: string, format: uuid }
        role: { type: string, enum: [viewer, editor, owner] }
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string, minLength: 1, maxLength: 128 }
        message: { type: string, minLength: 1, maxLength: 512 }
        current_revision: { type: integer, minimum: 0 }
        current_map_hash: { $ref: '#/components/schemas/SHA256' }
...